stt-backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.35.0
httpx==0.28.1
numpy==2.1.3
webrtcvad-wheels==2.0.14
faster-whisper==1.1.1
requests==2.32.5
python-dotenv==1.0.1
pytest==8.4.1
pytest-asyncio==0.24.0
